Remove unneeded cast.
As detected by Eclipse, and the test case doesn't depend on the explicit
casting.

+++ jadx-core/src/main/java/jadx/core/xmlgen/entry/ValuesParser.java
@@ -215,7 +215,7 @@ public class ValuesParser extends ParserConstants {
 	}
 
 	private static String floatToString(float value) {
-		return doubleToString((double) value);
+		return doubleToString(value);
 	}
 
 	public static Map<Integer, String> getAndroidResMap() {
